Move loadCrawlResults to a util lib
